time-llm_test/old/requirements.txt
# requirements.txt
"""
pandas==2.1.0
numpy==1.24.0
requests==2.31.0
beautifulsoup4==4.12.0
FinanceDataReader==0.9.50
pykrx==1.3.0
yfinance==0.2.28
langchain-community==0.0.38
python-dotenv==1.0.0
tqdm==4.66.0
scikit-learn==1.3.0
matplotlib==3.7.0
"""

# 설치 명령어
# pip install -r requirements.txt
